Creación de interfaz Cardiovascular metrics, cambio de métodos y tipos

+++ src/main/java/healthcalc/Adapter/HealthHospital.java
@@ -2,7 +2,7 @@ package healthcalc.Adapter;
 
 public interface HealthHospital {
 	
-	public float idealWeightA(char gender, float height) throws Exception;
+	public double idealWeightA(char gender, float height) throws Exception;
 	
 	public float bmrA(char gender, int age, float height, int weight) throws Exception;
